Rename query_web_app to queries_web
